requirements.txt
cic-eth-registry~=0.6.9
celery~=4.4.7
chainlib-eth~=0.3.0
urlybird~=0.0.2
